+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.1.2] - Unreleased
 - Encode granular features with base64
 - Refactor result format to generic ISCC data model
+- Add optional gradio GUI demo
 
 ## [0.1.1] - 2024-06-25
 - Handle text decoding errors gracefully